Rename `DomainSlider` to `DomainWidget`
