+++ apps/storybook/src/Customization.mdx
@@ -146,7 +146,7 @@ as you see fit. For instance, if you'd like to change the color of the curve of 
 | `--h5w-selector-option-focus--outlineColor` | `gray`                                       | Outline color of focused option      |
 | `--h5w-selector-groupLabel--color`          | `gray`                                       | Text color of group labels in menu   |
 
-##### Domain slider
+##### Domain widget
 
 | Name                                                 | Default                    | Description                                    |
 | ---------------------------------------------------- | -------------------------- | ---------------------------------------------- |
