+++ packages/app/src/vis-packs/core/complex/ComplexToolbar.tsx
@@ -1,7 +1,7 @@
 import type { Domain } from '@h5web/lib';
 import {
   ColorMapSelector,
-  DomainSlider,
+  DomainWidget,
   ScaleSelector,
   Selector,
   Separator,
@@ -44,7 +44,7 @@ function ComplexToolbar(props: Props) {
 
   return (
     <Toolbar interactions={getImageInteractions(keepRatio)}>
-      <DomainSlider
+      <DomainWidget
         dataDomain={dataDomain}
         customDomain={customDomain}
         scaleType={scaleType}
