+++ packages/lib/src/toolbar/controls/DomainWidget/Thumb.module.css
