+++ packages/lib/src/toolbar/controls/DomainWidget/Thumb.tsx
